Test algos and verified their functionality refactored Fit Algo refactored memory rendering sorted lists converted processes to useing longs

+++ src/main/java/model/Algos/Algo.java
@@ -4,7 +4,7 @@ import model.Process;
 
 public interface Algo {
 
-    public Integer allocPs(Process unallocated);
+    public Long allocPs(Process unallocated);
     public boolean deallocate(Process allocated);
     public String getName();
 }
